requirements.txt
python-dotenv==0.13.0
telegram-text==0.1.1
telethon==1.36.0
pytest-asyncio==0.23.8
